FIX - Chip Visibility and movement
- Fixed very small chips not displaying a tilemap
- Added a change to alter the movement speed for each chip
- (new issue) tiles don't act as a container for each other yet. Before,
they moved together because they moved the same speeds. Now since that's
changed you'll need to do something different

+++ ProjectMicro Solution/ProjectMicro/sprite.h
@@ -12,7 +12,8 @@ class Sprite
 		double x, y, midX, midY, oldX, oldY, vx, vy,
 			  speed, friction, maxSpeed,
 			  oldMaxSpeed;
-		int width, height, orgWidth, orgHeight;
+		double width, height, orgWidth, orgHeight;
+		//int width, height, orgWidth, orgHeight;
 		SDL_Rect collisionRect;
 		SDL_Rect collisionVert;
 		SDL_Rect collisionHorz;
